Split JSON and Table formatters into different files
