+++ internal/formatter/values.go
@@ -5,6 +5,8 @@ import (
 	"time"
 )
 
+const MaxLength = 50
+
 func FormatValue(value interface{}) string {
 	switch data := value.(type) {
 	case string:
@@ -21,12 +23,12 @@ func formatString(value string) string {
 		return date.Format("2006-01-02 15:04:05")
 	}
 
-	return truncate(value, 50)
+	return truncate(value)
 }
 
-func truncate(input string, maxLength int) string {
-	if len(input) > maxLength {
-		return input[:maxLength] + "..."
+func truncate(input string) string {
+	if len(input) > MaxLength {
+		return input[:MaxLength] + "..."
 	}
 	return input
 }
